requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
alembic==1.14.0
aiosqlite==0.20.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Validation
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0

# Hardware tools
pyserial==3.5
hidapi==0.14.0

# Utilities
python-dotenv==1.0.1
loguru==0.7.3
